# services/licensing-service/openapi.yaml
openapi: 3.1.0
info:
  title: HDICR Licensing Service API
  version: 3.0.0
  summary: Licensing request lifecycle and actor licensing management contract.
  description: >-
    Defines endpoint contracts for submitting license requests, making actor decisions,
    resolving actor licensing context, and retrieving license/request aggregates.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: https://api.trulyimagined.com
    description: Production API
...
security:
  - bearerAuth: []
paths:
  /v1/license/request:
    post:
      tags: [Licensing]
      operationId: requestLicense
      summary: Create a license request for an actor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LicenseRequestCreateInput'
      responses:
        '201':
          description: License request created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LicenseRequestCreateResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/license/actor/{actorId}:
    get:
      tags: [Licensing]
      operationId: listLicenseRequestsForActor
      summary: List license requests targeting a specific actor
      parameters:
        - $ref: '#/components/parameters/ActorIdPath'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Actor request listing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActorRequestsResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Auth0 RS256-signed JWT. Required on all endpoints (except OPTIONS
        preflight). Scope enforcement is applied after signature verification.
    hdicrOAuth2:
      type: oauth2
      description: >
        OAuth 2.1 client credentials used by TI or any authorised HDICR client.
        Tokens are acquired via POST /oauth/token with grant_type=client_credentials.
        The required scope per endpoint determines which operations are permitted.
      flows:
        clientCredentials:
          tokenUrl: https://{auth0Domain}/oauth/token
          scopes:
            hdicr:licensing:read: Read licensing requests, decisions, and actor license data
            hdicr:licensing:write: Create, approve, or reject licensing requests
  parameters:
    ActorIdPath:
      name: actorId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
    Offset:
      name: offset
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
  responses:
    BadRequest:
      description: Invalid request payload or query parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Authenticated caller lacks required scope or role
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Requested resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalServerError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
      additionalProperties: true
    LicenseStatus:
      type: string
      enum: [pending, approved, rejected, expired]
    CurrencyCode:
      type: string
      minLength: 3
      maxLength: 3
    CompensationAmount:
      type: [number, 'null']
      minimum: 0
    LicenseRequestSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        actorId:
          type: string
          format: uuid
        requesterName:
          type: string
        requesterEmail:
          type: string
          format: email
        requesterOrganization:
          type: [string, 'null']
        projectName:
          type: string
        projectDescription:
          type: string
        usageType:
          type: string
        intendedUse:
          type: string
        compensationOffered:
          $ref: '#/components/schemas/CompensationAmount'
        compensationCurrency:
          $ref: '#/components/schemas/CurrencyCode'
        status:
          $ref: '#/components/schemas/LicenseStatus'
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - actorId
        - requesterName
        - requesterEmail
        - projectName
        - projectDescription
        - usageType
        - intendedUse
        - status
        - createdAt
      additionalProperties: false
    LicenseRequestCreateInput:
      type: object
      properties:
        actorId:
          type: string
          format: uuid
        requesterName:
          type: string
        requesterEmail:
          type: string
          format: email
        requesterOrganization:
          type: string
        projectName:
          type: string
        projectDescription:
          type: string
        usageType:
          type: string
        intendedUse:
          type: string
        durationStart:
          type: string
          format: date-time
        durationEnd:
          type: string
          format: date-time
        compensationOffered:
          $ref: '#/components/schemas/CompensationAmount'
        compensationCurrency:
          $ref: '#/components/schemas/CurrencyCode'
      required:
        - actorId
        - requesterName
        - requesterEmail
        - projectName
        - projectDescription
        - usageType
        - intendedUse
      additionalProperties: false
    LicenseRequestCreateResponse:
      type: object
      properties:
        success:
          type: boolean
          const: true
        message:
          type: string
        request:
          type: object
          properties:
            id:
              type: string
              format: uuid
            actorId:
              type: string
              format: uuid
            projectName:
              type: string
            usageType:
              type: string
            status:
              $ref: '#/components/schemas/LicenseStatus'
            createdAt:
              type: string
              format: date-time
          required: [id, actorId, projectName, usageType, status, createdAt]
          additionalProperties: false
      required: [success, message, request]
      additionalProperties: false
    ActorRequestsResponse:
      type: object
      properties:
        actorId:
          type: string
          format: uuid
        requests:
          type: array
          items:
            $ref: '#/components/schemas/LicenseRequestSummary'
        pagination:
          type: object
          properties:
            limit:
              type: integer
            offset:
              type: integer
            total:
              type: integer
          required: [limit, offset, total]
          additionalProperties: false
      required: [actorId, requests, pagination]
      additionalProperties: false
...